+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed integer coercion when reading variables in GraphQL API
+
 ## [1.0.1] 2023-09-06
 
 ### Fixed